Add Drawer Kick (#2)
The official spec for Receiptline omits any functions for the drawer
kick function on printers. Instead they suggest the use of the
`{command:}` option. I'm not a fan as this will always be
language-specific, and for this particular library I have the leeway to
extend the syntax of options at least a little bit.

This adds a `{drawer:kick}` option to the parser to allow documents to
indicate the cash drawer kick should be, well, kicked. The alias value
`kick` maps to `1`, for output 1. Some printers support two outputs, so
`2` is also a supported input value.

* `{drawer:kick}`
* `{drawer:1}`
* `{drawer:2}`

are valid options for syntax and will result in the drawer kick command
being added to the document.