Backport structural changes from WebZLP (#15)
In prep for v2.0 of the [WebZLP](https://github.com/cellivar/webzlp)
library I've made significant alterations to the structure of that
project. The goal was to improve the modularity of the code and promote
more sustainable patterns based on lessons learned writing this library.

The primary focus has been on making the library more async-friendly,
and generally easier to integrate within UI frameworks such as React.

These changes in this PR are mostly about shuffling things around
instead of visible changes, and the effects on actual consumers are
fairly limited. For an example of such changes see the demo.html page
for the changes you'll need to make.